requirements.txt
# Google
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
oauth2client

# Debuging
ipython

#Environment
environs